+++ geodepy/tests/resources/Site13-134.fbk
@@ -1,0 +1,217 @@
+HORIZ ANGLE RIGHT
+VERT ANGLE ZENITH
+UNITS FOOT DMS
+JOB Site13-134
+UNITS METER DMS
+! SurvPC Version 5.08.1
+! CRD: Alphanumeric
+! User Defined: AUSTRALIA/GDA94/MGA zone 56 E150-156
+! Equipment:   Geomax Robotic,  Zoom 90
+! TS Scale: 1.00000000
+! Scale Point not used
+! EDM Mode: Standard
+NEZ 1302 7001228.5600 293989.1880 397.5280 "Bs"
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+NEZ 13 7001264.0430 294137.7230 394.0080
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+PRISM 2.108
+STN 13 1.652
+BS 1302 256.3353
+F1 VA 1302 256.3353 152.758 88.3029 "Bs"
+! Calculated: AR256°33'53", HD152.714, Z397.528
+! Measured: AR256°33'53", HD152.706, Z397.529
+! Delta: AR0°00'00", HD-0.008, Z0.001
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+! Set Collection with Obs Order 123...321...
+STN 13 1.652
+BS 1302 256.3353
+! DT05-14-2018
+! TM15:37:39
+! DT05-14-2018
+! TM15:37:41
+F1 VA 1302 256.3353 152.758 88.3030 "Bs"
+PRISM 1.827
+F1 VA 1304 358.3406 40.973 88.4301 "Fs"
+F2 VA 1304 178.3401 40.973 271.1651 "Fs"
+PRISM 2.108
+F2 VA 1302 76.3353 152.758 271.2929 "Bs"
+F1 VA 1302 256.3353 152.758 88.3030 "Bs"
+PRISM 1.767
+F1 VA 1305 357.1807 45.316 88.5033 "Fs"
+F2 VA 1305 177.1804 45.317 271.0921 "Fs"
+PRISM 2.108
+F2 VA 1302 76.3353 152.758 271.2929 "Bs"
+F1 VA 1302 256.3352 152.758 88.3028 "Bs"
+PRISM 1.827
+F1 VA 1304 358.3403 40.973 88.4300 "Fs"
+F2 VA 1304 178.3405 40.973 271.1651 "Fs"
+PRISM 2.108
+F2 VA 1302 76.3351 152.757 271.2929 "Bs"
+F1 VA 1302 256.3352 152.758 88.3028 "Bs"
+PRISM 1.767
+F1 VA 1305 357.1803 45.317 88.5033 "Fs"
+F2 VA 1305 177.1800 45.316 271.0920 "Fs"
+PRISM 2.108
+F2 VA 1302 76.3351 152.757 271.2929 "Bs"
+F1 VA 1302 256.3352 152.739 88.3027 "Bs"
+PRISM 1.827
+F1 VA 1304 358.3407 40.973 88.4300 "Fs"
+F2 VA 1304 178.3402 40.973 271.1651 "Fs"
+PRISM 2.108
+F2 VA 1302 76.3352 152.758 271.2930 "Bs"
+! SS,OP13,FP1302,AR256.3353,ZE88.3029,SD152.754774,--Bs
+! LS,HI1.652000,HR1.827000
+! SS,OP13,FP1304,AR358.3404,ZE88.4305,SD40.973038,--Fs
+! LS,HI1.652000,HR2.108000
+F1 VA 1302 256.3352 152.739 88.3027 "Bs"
+PRISM 1.767
+F1 VA 1305 357.1801 45.316 88.5033 "Fs"
+F2 VA 1305 177.1759 45.316 271.0921 "Fs"
+PRISM 2.108
+F2 VA 1302 76.3352 152.758 271.2930 "Bs"
+! SS,OP13,FP1302,AR256.3353,ZE88.3029,SD152.754774,--Bs
+! LS,HI1.652000,HR1.767000
+! SS,OP13,FP1305,AR357.1803,ZE88.5036,SD45.316350,--Fs
+! LS,HI1.652000,HR2.108000
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+STN 13 1.652
+BS 1302 256.3352
+F1 VA 1306 256.3353 152.760 88.3019 "chk1302 from 13"
+! DT05-14-2018
+! TM15:45:49
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+PRISM 1.599
+STN 1304 1.827
+BS 13 178.3404
+F1 VA 13 178.3404 40.975 91.2107 ""
+! Calculated: AR178°34'04", HD40.963, Z394.008
+! Measured: AR178°34'04", HD40.963, Z394.011
+! Delta: AR0°00'00", HD0.000, Z0.003
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+! P.C. mm Applied: 0.0000 (Reflectorless:foresight)
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+PRISM 1.767
+F1 VA 1307 345.3311 4.446 90.0141 "Chk1305"
+! DT05-14-2018
+! TM15:57:54
+! P.C. mm Applied: 0.0000 (Reflectorless:foresight)
+PRISM 0.000
+F1 VA 1308 7.2824 3.557 99.3411 "Cra"
+! DT05-14-2018
+! TM16:00:55
+F1 VA 1309 37.3842 3.598 82.4158 "Crb"
+! DT05-14-2018
+! TM16:02:33
+F1 VA 1310 34.4111 1.833 108.5758 "crc"
+! DT05-14-2018
+! TM16:04:23
+F1 VA 1311 31.5612 3.320 106.3933 "Crd"
+! DT05-14-2018
+! TM16:06:26
+F1 VA 1312 32.2323 3.321 105.0110 "Cre-100 above crd"
+! DT05-14-2018
+! TM16:07:37
+F1 VA 1313 8.2821 3.995 110.5909 "Sla"
+! DT05-14-2018
+! TM16:08:26
+F1 VA 1314 57.0101 3.337 115.3615 "Slc"
+! DT05-14-2018
+! TM16:10:34
+F1 VA 1315 17.1005 2.290 128.5220 "Sld"
+! DT05-14-2018
+! TM16:11:25
+F1 VA 1316 30.0228 3.305 111.2112 "Bolt"
+! DT05-14-2018
+! TM16:13:01
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+! Backsight Check: PT 13
+! Calculated: AR178°34'04", HD40.963, Z394.008
+! Measured: AR178°34'04", HD40.963, Z394.010
+! Delta: AR0°00'01", HD0.000, Z0.002
+! P.C. mm Applied: 0.0000 (Reflectorless:foresight)
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+PRISM 1.599
+STN 1305 1.767
+BS 13 177.1803
+F1 VA 13 177.1803 45.318 91.1311 ""
+! Calculated: AR177°18'03", HD45.307, Z394.008
+! Measured: AR177°18'03", HD45.307, Z394.011
+! Delta: AR0°00'00", HD0.000, Z0.003
+! P.C. mm Applied: 0.0000 (Reflectorless:foresight)
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+! P.C. mm Applied: 0.0000 (Reflectorless:foresight)
+PRISM 0.000
+F1 VA 1317 117.5659 1.871 108.2204 "Cra"
+! DT05-14-2018
+! TM16:22:30
+F1 VA 1318 114.1430 3.637 82.4548 "Crb"
+! DT05-14-2018
+! TM16:23:54
+F1 VA 1319 144.0140 3.614 99.2633 "crc"
+! DT05-14-2018
+! TM16:24:46
+F1 VA 1320 119.5255 3.357 106.2536 "crc"
+! DT05-14-2018
+! TM16:28:48
+F1 VA 1320 119.5254 3.357 106.2533 "Crd"
+! DT05-14-2018
+! TM16:29:18
+F1 VA 1321 119.2749 3.360 104.4807 "Cre-100 above crd"
+! DT05-14-2018
+! TM16:30:27
+F1 VA 1322 110.3959 2.280 128.4928 "Sla"
+! DT05-14-2018
+! TM16:31:17
+F1 VA 1323 100.3322 3.996 111.0905 "Slb"
+! DT05-14-2018
+! TM16:32:01
+F1 VA 1324 126.1902 4.733 107.4155 "Slc"
+! DT05-14-2018
+! TM16:33:00
+F1 VA 1325 148.0138 3.402 114.4804 "Sld"
+! DT05-14-2018
+! TM16:33:49
+F1 VA 1326 121.4749 3.343 111.0323 "Bolt"
+! DT05-14-2018
+! TM16:34:38
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+! Backsight Check: PT 13
+! Calculated: AR177°18'03", HD45.307, Z394.008
+! Measured: AR177°18'18", HD45.308, Z394.010
+! Delta: AR0°00'15", HD0.001, Z0.002
+! P.C. mm Applied: 0.0000 (Reflectorless:foresight)
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+PRISM 2.111
+F1 VA 1327 241.0732 167.211 88.5650 "chk1302"
+! DT05-14-2018
+! TM16:41:51
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+! Backsight Check: PT 13
+! Calculated: AR177°18'03", HD45.307, Z394.008
+! Measured: AR177°18'10", HD45.308, Z394.011
+! Delta: AR0°00'07", HD0.000, Z0.003
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
+F1 VA 1328 241.0733 167.211 88.5648 "chk1302"
+! DT05-14-2018
+! TM16:43:30
+! P.C. mm Applied: -11.3000 (Leica 360:backsight)
+! Backsight Check: PT 13
+! Calculated: AR177°18'03", HD45.307, Z394.008
+! Measured: AR177°18'10", HD45.307, Z394.011
+! Delta: AR0°00'07", HD0.000, Z0.003
+! P.C. mm Applied: -30.0000 (-30mm:foresight)
