+++ geodepy/tests/resources/Site13-134.stn
@@ -1,0 +1,11 @@
+!#=DNA 3.01 STN    14.05.2018         GDA94                       10
+SB13                CCC UTM   294138.3249      7001265.4623        393.8644            S56 
+1302                FFF UTM   293989.7899      7001229.9793        397.3844            S56 Bs
+1304                FFF UTM   294137.3009      7001306.4123        394.6064            S56 Fs
+1305                FFF UTM   294136.1909      7001310.7193        394.6644            S56 Fs
+1308                FFF UTM   294137.7569      7001309.8903        395.8414            S56 Cra
+1309                FFF UTM   294139.4809      7001309.2383        396.8904            S56 Crb
+1310                FFF UTM   294138.2879      7001307.8383        395.8374            S56 crc
+1311                FFF UTM   294138.9839      7001309.1123        395.4814            S56 Crd
+1312                FFF UTM   294139.0199      7001309.1213        395.5724            S56 Cre-100 above crd
+1316                FFF UTM   294138.8419      7001309.0763        395.2294            S56 Bolt
