+++ geodepy/tests/resources/Site13-134.txt
@@ -1,0 +1,27 @@
+13,294137.723,7001264.043,394.008,
+1302,293989.188,7001228.560,397.528,Bs
+1304,294136.699,7001304.993,394.750,Fs,DATE:05-14-2018,TIME:15:37:39
+1305,294135.589,7001309.300,394.808,Fs,DATE:05-14-2018,TIME:15:37:41
+1306,293989.194,7001228.563,397.534,chk1302 from 13,DATE:05-14-2018,TIME:15:45:49
+1307,294135.590,7001309.299,394.808,Chk1305,DATE:05-14-2018,TIME:15:57:54
+1308,294137.155,7001308.471,395.985,Cra,DATE:05-14-2018,TIME:16:00:55
+1309,294138.879,7001307.819,397.034,Crb,DATE:05-14-2018,TIME:16:02:33
+1310,294137.686,7001306.419,395.981,crc,DATE:05-14-2018,TIME:16:04:23
+1311,294138.382,7001307.693,395.625,Crd,DATE:05-14-2018,TIME:16:06:26
+1312,294138.418,7001307.702,395.716,Cre-100 above crd,DATE:05-14-2018,TIME:16:07:37
+1313,294137.249,7001308.683,395.146,Sla,DATE:05-14-2018,TIME:16:08:26
+1314,294139.224,7001306.631,395.135,Slc,DATE:05-14-2018,TIME:16:10:34
+1315,294137.225,7001306.696,395.140,Sld,DATE:05-14-2018,TIME:16:11:25
+1316,294138.240,7001307.657,395.373,Bolt,DATE:05-14-2018,TIME:16:13:01
+1317,294137.158,7001308.468,395.985,Cra,DATE:05-14-2018,TIME:16:22:30
+1318,294138.879,7001307.818,397.033,Crb,DATE:05-14-2018,TIME:16:23:54
+1319,294137.684,7001306.415,395.982,crc,DATE:05-14-2018,TIME:16:24:46
+1320,294138.381,7001307.696,395.625,Crd,DATE:05-14-2018,TIME:16:29:18
+1321,294138.418,7001307.702,395.716,Cre-100 above crd,DATE:05-14-2018,TIME:16:30:27
+1322,294137.251,7001308.673,395.145,Sla,DATE:05-14-2018,TIME:16:31:17
+1323,294139.253,7001308.617,395.133,Slb,DATE:05-14-2018,TIME:16:32:01
+1324,294139.222,7001306.629,395.136,Slc,DATE:05-14-2018,TIME:16:33:00
+1325,294137.225,7001306.680,395.148,Sld,DATE:05-14-2018,TIME:16:33:49
+1326,294138.241,7001307.656,395.374,Bolt,DATE:05-14-2018,TIME:16:34:38
+1327,293989.191,7001228.568,397.536,chk1302,DATE:05-14-2018,TIME:16:41:51
+1328,293989.191,7001228.569,397.538,chk1302,DATE:05-14-2018,TIME:16:43:30
